updated header changes after pr review
